created a reusable promise based confirmationmodal
This is wizardry!

+++ frontend/src/components/pages/Rubric_component/dependencies/MarkingRangeTableRow.jsx
@@ -7,10 +7,37 @@ import BASE_API_URL from '../../../../BASE_API_URL'
 import axiosToBackend from '../../../../axiosToBackend'
 
 import DoubleClickModifyCell from './DoubleClickModifyCell'
+import useConfirmation from '../../../hooks/useConfirmation'
 
 const MarkingRangeTableRow = ({ rubricComponent, setRubricComponent, index }) => {
+
+    const [confirm, ConfirmationModal] = useConfirmation()
+
+
+
+    const DeletionModalBody = () => {
+
+
+        return (<>
+            <div className='mb-0'>Are you sure you want to delete the following rating range:</div>
+            <strong className='mt-1'>{rubricComponent.rating_ranges[index].rating_desc}</strong>
+            <p>This action is irreversible!</p>
+        </>
+        )
+    }
+
+
     async function handleRemoveButtonClick(event) {
+
+
+
         const ratingRangeIdToDelete = rubricComponent.rating_ranges[index].rating_range_id
+        console.log(rubricComponent.rating_ranges[index])
+        const confirmation = await confirm(<DeletionModalBody />)
+        console.log(confirmation)
+        if (!confirmation) {
+            return
+        }
 
         const apiDeleteUrl = `${BASE_API_URL}module/${rubricComponent.module_id}/exam/${rubricComponent.exam_id}/rubric/${rubricComponent.rubric_component_id}/rating_range/${ratingRangeIdToDelete}`
         const responseFromDelete = await axiosToBackend.delete(apiDeleteUrl)
@@ -18,7 +45,7 @@ const MarkingRangeTableRow = ({ rubricComponent, setRubricComponent, index }) =>
         if (responseFromDelete.status === 204) {
             // do stuff
             let updatedRubricComponent = { ...rubricComponent }
-            let updatedRatingRanges = rubricComponent.rating_ranges.slice(0, index).concat(rubricComponent.rating_ranges.slice(index+1))
+            let updatedRatingRanges = rubricComponent.rating_ranges.slice(0, index).concat(rubricComponent.rating_ranges.slice(index + 1))
             updatedRubricComponent.rating_ranges = updatedRatingRanges
             setRubricComponent(updatedRubricComponent)
         } else {
@@ -29,31 +56,35 @@ const MarkingRangeTableRow = ({ rubricComponent, setRubricComponent, index }) =>
 
 
     return (
-        <tr >
-            <DoubleClickModifyCell
-                parameterInCell={'rating_desc'}
-                rubricComponent={rubricComponent}
-                setRubricComponent={setRubricComponent}
-                index={index}
-            />
-            <DoubleClickModifyCell
-                parameterInCell={'rating_min_incl'}
-                rubricComponent={rubricComponent}
-                setRubricComponent={setRubricComponent}
-                index={index}
-            />
-            <DoubleClickModifyCell
-                parameterInCell={'rating_max_incl'}
-                rubricComponent={rubricComponent}
-                setRubricComponent={setRubricComponent}
-                index={index}
-            />
-            <td>
-                <Button variant="warning" onClick={handleRemoveButtonClick}>
-                    Remove
-                </Button>
-            </td>
-        </tr>
+
+        <>
+            <tr >
+                <DoubleClickModifyCell
+                    parameterInCell={'rating_desc'}
+                    rubricComponent={rubricComponent}
+                    setRubricComponent={setRubricComponent}
+                    index={index}
+                />
+                <DoubleClickModifyCell
+                    parameterInCell={'rating_min_incl'}
+                    rubricComponent={rubricComponent}
+                    setRubricComponent={setRubricComponent}
+                    index={index}
+                />
+                <DoubleClickModifyCell
+                    parameterInCell={'rating_max_incl'}
+                    rubricComponent={rubricComponent}
+                    setRubricComponent={setRubricComponent}
+                    index={index}
+                />
+                <td>
+                    <Button variant="warning" onClick={handleRemoveButtonClick}>
+                        Remove
+                    </Button>
+                </td>
+            </tr>
+            <ConfirmationModal />
+        </>
     )
